move ConstraintRemover to task to solve multitasking related issues
